Add crf to VideoEncoder API (#1031)

+++ src/torchcodec/_core/StreamOptions.h
@@ -47,7 +47,7 @@ struct VideoStreamOptions {
   // Encoding options
   // TODO-VideoEncoder: Consider adding other optional fields here
   // (bit rate, gop size, max b frames, preset)
-  std::optional<int> crf;
+  std::optional<double> crf;
 
   // Optional pixel format for video encoding (e.g., "yuv420p", "yuv444p")
   // If not specified, uses codec's default format.
